refactor range filtering
