+++ fold_node/src/schema/types/field/mod.rs
@@ -2,10 +2,12 @@ pub mod common;
 pub mod single_field;
 pub mod collection_field;
 pub mod range_field;
+pub mod range_filter;
 pub mod variant;
 
 pub use common::{Field, FieldCommon, FieldType};
 pub use single_field::SingleField;
 pub use collection_field::CollectionField;
-pub use range_field::{RangeField, RangeFilter, RangeFilterResult};
+pub use range_field::RangeField;
+pub use range_filter::{RangeFilter, RangeFilterResult};
 pub use variant::FieldVariant;
